requirements.txt
fastapi
uvicorn[standard]
kerykeion
python-dotenv
